fix: address even more feedback
