+++ packages/core/src/index.tsx
@@ -2,6 +2,7 @@ export { BaseNavigationContainer } from './BaseNavigationContainer';
 export { createNavigationContainerRef } from './createNavigationContainerRef';
 export { createNavigatorFactory } from './createNavigatorFactory';
 export { CurrentRenderContext } from './CurrentRenderContext';
+export { UNSTABLE_getLoaderForState } from './DataLoading';
 export { findFocusedRoute } from './findFocusedRoute';
 export { getActionFromState } from './getActionFromState';
 export { getFocusedRouteNameFromRoute } from './getFocusedRouteNameFromRoute';
@@ -28,7 +29,6 @@ export {
   type StaticScreenConfigLinking,
   type StaticScreenConfigScreen,
   type StaticScreenProps,
-  UNSTABLE_getLoaderForState,
 } from './StaticNavigation';
 export { ThemeContext } from './theming/ThemeContext';
 export { ThemeProvider } from './theming/ThemeProvider';
